+++ CHANGELOG.md
@@ -13,6 +13,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+## [1.131.0]
+
+### Added
+- Save theme in user settings
+
+### Fixed
+- multiline does not work correctly with string values
+
 ## [1.130.0]
 
 ### Added